# Conflicts:
#	src/main/java/com/example/homeGym/instructor/dto/CommentDto.java
#	src/main/java/com/example/homeGym/instructor/service/InstructorService.java
#	src/main/java/com/example/homeGym/instructor/service/UserProgramService.java

+++ src/main/java/com/example/homeGym/instructor/controller/ProgramController.java
@@ -2,6 +2,7 @@ package com.example.homeGym.instructor.controller;
 
 import com.example.homeGym.instructor.dto.ProgramDto;
 import com.example.homeGym.instructor.service.ProgramService;
+import com.example.homeGym.instructor.service.UserProgramService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -16,6 +17,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ProgramController {
     private final ProgramService programService;
+    private final UserProgramService userProgramService;
 
     @GetMapping()
     public String createPage(
@@ -27,20 +29,10 @@ public class ProgramController {
 
     @PostMapping()
     public String requestCreate(
-            @Valid @ModelAttribute ProgramDto programDto,
-            BindingResult bindingResult,
-            Model model
+            @ModelAttribute ProgramDto programDto
     ) {
-        if (bindingResult.hasErrors()) {
-            return "/instructor/program";
-        }
-        try {
-            programService.createProgram(programDto);
-            return "redirect:/instructor/program";
-        } catch (Exception e) {
-            model.addAttribute("errorMessage", e.getMessage());
-            return "error-page"; // 에러 페이지로 리다이렉트
-        }
+        programService.createProgram(programDto);
+        return "redirect:/instructor/program";
     }
 
 
@@ -66,12 +58,13 @@ public class ProgramController {
         }
         try {
             programService.updateProgram(programId, programDto);
-            return "redirect:/program";
+            return "redirect:/instructor/program";
         } catch (Exception e) {
             model.addAttribute("errorMessage", e.getMessage());
             return "error-page"; // 에러 페이지로 리다이렉트
         }
     }
+
 
     @DeleteMapping("/{programId}")
     public String deleteProgram(
@@ -79,11 +72,18 @@ public class ProgramController {
             Model model
     ) {
         try {
+            // 프로그램 삭제 시 관련 수강생 확인
+            if (userProgramService.hasEnrolledUsers(programId)) {
+                throw new RuntimeException("Cannot delete program with enrolled users");
+            }
+
             programService.deleteProgram(programId);
-            return "redirect:/program";
+            return "redirect:/instructor/program";
         } catch (Exception e) {
             model.addAttribute("errorMessage", e.getMessage());
             return "error-page"; // 에러 페이지로 리다이렉트
         }
     }
+
+
 }
